move ThumbnailGridErrors into its own file
